Make the algorithm work in the modeler setting the extent dynamically
When using model input (e.g. vector layer) as source of the extent,
the QGIS processing framework sets the string value (the one available in
`parameters['extent']` to the layer id. However, the code in this plugin
assumes it will be a list of coordinates as it shown when using manually.
Luckily, the processing framework deals the layer id values correctly
when you pass them to `parameterAsExtentGeometry` and
`parameterAsExtentCrs`.